requirements.txt
Django<2.2
django-allauth>=0.38.0
django-filter>=2.1.0
django-widget-tweaks>=1.4.3
djangorestframework>=3.9.1
django-extensions>=2.1.4
django-werkzeug-debugger-runserver>=0.3.1
gevent>=1.4.0
gevent-websocket>=0.10.1
lxml>=4.3.0
Markdown>=3.0.1
oauthlib>=3.0.1
Pillow>=5.3.0
pandas>=0.23.4
pydot>=1.3.0
pyquery>=1.4.0
pyzmq>=17.1.2
qrcode>=6.1
requests>=2.20.1
requests-oauthlib>=1.2.0
six>=1.11.0
setuptools>=18.0.0
scipy>=1.1.0
tensorboard<2.13.0
tensorflow<2.13.1
Keras>=2.2.2
Keras-Applications>=1.0.4
Keras-Preprocessing>=1.0.2
h5py>=2.8.0
numpy==1.16.4
selenium


